+++ config/routes.rb
@@ -1,6 +1,9 @@
 WwwSleepdataOrg::Application.routes.draw do
 
   resources :topics, path: "forum" do
+    member do
+      post :admin
+    end
     resources :comments do
       collection do
         post :preview