+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Added the ability for registered users to create topics on the forum and to post comments on other topics
   - Core and AUG members are now highlighted as such on their forum posts
   - Comments can be previewed before being posted to the forum topic
+  - System admins can lock, pin, and delete topics
 
 ## 0.8.1 (May 28, 2014)